+++ CHANGELOG.md
@@ -1,4 +1,16 @@
+## 1.0.0
+
+* **BREAKING** - Change builder methods to expose `CalendarAdvancedController`
+* Expose default cell builder
+* Allow to select today by default when calling `goToToday()`
+* Change calendar arrows position
+* Avoid cell selection if date is not in range
+* Added swipe gestures to change modes or goBack/goforward in calendar
+* Added parameters to enable or disable gestures and to enable or disable highlight initialDate
+* Bugfix
+
 ## 0.0.2
+
 Added docs and improved Readme.
 
 ## 0.0.1